adding customer and shoppingCart

+++ FoodOrderingAPI/Controllers/AddressController.cs
@@ -1,11 +1,14 @@
 ﻿using FoodOrderingAPI.DTO;
 using FoodOrderingAPI.Interfaces;
 using FoodOrderingAPI.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 
 namespace FoodOrderingAPI.Controllers
 {
+    [Authorize(Roles = "Customer")]
     [Route("api/[controller]")]
     [ApiController]
     public class AddressController : ControllerBase
@@ -17,6 +20,11 @@ namespace FoodOrderingAPI.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAll(string UserName)
         {
+            var userNameClaim = User.FindFirstValue(ClaimTypes.Name);
+            if (userNameClaim != UserName)
+            {
+                return Forbid("You are not authorized to View this Customer's Address.");
+            }
             return Ok(await addressRepo.GetAllAddresses(UserName));
         }
         [HttpGet("ById")]
@@ -30,6 +38,11 @@ namespace FoodOrderingAPI.Controllers
         [HttpPost]
         public async Task<IActionResult> Add(string userName, AddressDTO address)
         {
+            var userNameClaim = User.FindFirstValue(ClaimTypes.Name);
+            if (userNameClaim != userName)
+            {
+                return Forbid("You are not authorized to add adddress to this Customer's Address.");
+            }
             if (ModelState.IsValid)
             {
                 await addressRepo.Add(userName, address);
@@ -43,11 +56,19 @@ namespace FoodOrderingAPI.Controllers
 
         }
         [HttpPut]
-        public async Task<IActionResult> Update(Guid AddressId, AddressDTO address)
+        public async Task<IActionResult> Update(Guid AddressId, AddressDTO addressdto)
         {
+            Address address= await addressRepo.GetAddress(AddressId);
+            if (address ==null)
+                return NotFound();
+            var userIdClaim = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            if (userIdClaim != address.CustomerID)
+            {
+                return Forbid("You are not authorized to Update this Customer's Address.");
+            }
             if (ModelState.IsValid)
             {
-                if (await addressRepo.Update(AddressId, address))
+                if (await addressRepo.Update(AddressId, addressdto))
                 {
                     await addressRepo.Save();
                     return Ok("add Address Successfully");
@@ -64,6 +85,14 @@ namespace FoodOrderingAPI.Controllers
         [HttpDelete]
         public async Task<IActionResult> Delete(Guid AddressId)
         {
+            Address address = await addressRepo.GetAddress(AddressId);
+            if (address == null)
+                return NotFound();
+            var userIdClaim = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            if (userIdClaim != address.CustomerID)
+            {
+                return Forbid("You are not authorized to Delete this Customer's Address.");
+            }
             if (await addressRepo.Delete(AddressId))
             {
                 await addressRepo.Save();
